Remove dependency to properties singleton in show resources test
